+++ CHANGELOG.md
@@ -21,6 +21,11 @@ about writing changes to this log.
   - Updated `CaseEvent` with `senders` and `recipients` property.
   - Removed `CaseEventPartyRelation`.
 - Added user signatures [TVIST1-726](https://jira.itkdev.dk/browse/TVIST1-726)
+- [TVIST1-436](https://jira.itkdev.dk/browse/TVIST1-436)
+  Implemented Next Generation Digitapl Post (NgDP) and handling of
+  Beskedfordeler messages.
+- [TVIST1-436](https://jira.itkdev.dk/browse/TVIST1-436)
+  Fixed CPR lookup.
 
 ## [1.1.2] 2023-02-17